adding test and inits to test data folders
